fix(coap): use consistent error type
The previous distinction in "this function parses and thus returns a
minicbor error" and "this function decrypts and therefore may return a
decryption failure" was already broken before the current additions, and
things have become worse. The new type covers the previous cases, and
can be reduced to a ZST if needed.