chore(coap): generalize lint exception
Fields in a derived CBOR map have a function even when they are never
read. The previous version of expect passed CI in some placed, but
triggered in others, because the interactions with non-CBOR derives (eg.
defmt) are not trivial, and not exhaustively checked (and only that
would make sure they are expressed right).